+++ hardhat.config.ts
@@ -109,7 +109,7 @@ const config: HardhatUserConfig = {
         },
       },
       {
-        version: '0.8.4',
+        version: '0.8.12',
         settings: {
           ...settings,
           metadata: {